build: move ts-node to be a dev tool only
My work ends up as a library, a service, a scheduled ETL or a cli tool.
So the work ends up packaged as a library or as a docker image.
In either case, I want compiled js in the npm package or docker image.
So ts-node is really only a dev tool.